Practiced 1.save a text file using console application 2.append logs of a running application 3. read a file on start up

+++ DAY4/DAY-4-GIT-EXERCISE/DAY-4-GIT-EXERCISE/Program.cs
@@ -36,7 +36,7 @@ namespace Basic_Calculator
                             Console.WriteLine("the multiplication is " + (num1 * num2));
                             break;
                         case 4:
-                            Console.WriteLine("the division is " + Convert.ToDouble(num1 /(double) num2));
+                            Console.WriteLine("the division is " + Convert.ToDouble(num1 / num2));
                             break;
 
                     }
